// calculator-fe/src/app/calculator/components/keypad/keypad.enum.ts
export enum KeypadOperator {
  Delete = 'Delete',
  Backspace = 'Backspace',
  Addition = '+',
  Subtraction = '-',
  Multiplication = '*',
  Division = '/',
  Equals = '=',
  DecimalPoint = '.',
}

export enum KeypadNumber {
  Zero = '0',
  One = '1',
  Two = '2',
  Three = '3',
  Four = '4',
  Five = '5',
  Six = '6',
  Seven = '7',
  Eight = '8',
  Nine = '9',
}